YOU_HURT_ME_1.0
This repository features a malware designed for working directly from your browser. It includes an End User License Agreement (EULA) that users must consent to before running the script. The script simulates the behavior of the classic "You Are An Idiot" virus and is intended for educational purposes only. Users are encouraged to review and agree to the EULA before using the script, and to use it responsibly and ethically. The author does not assume responsibility for any misuse or consequences resulting from the script's usage. Proceed with caution and adhere to ethical guidelines.